refactor(twin-gmail): unify domain modules and Gate-1 MCP tools
Fold rest-store into domain/*, split search parse/match with SQL prefilter,
promote MCP 10→13 (get_message + sensitive labels), and add FIDELITY.md.
