+++ packages/twin-gmail/fixtures/README.md
@@ -1,4 +1,4 @@
-# Gmail twin fixtures (Gate 0 oracles)
+# Gmail twin fixtures (Gate 0/1 oracles)
 
 Immutable upstream captures for `@pome-sh/twin-gmail`. Normal tests must not require Google credentials.
 
@@ -15,8 +15,8 @@ Immutable upstream captures for `@pome-sh/twin-gmail`. Normal tests must not req
 | File | Purpose |
 | --- | --- |
 | `mcp-tools-list.raw.json` | Live unauthenticated `tools/list` (13 tools as returned) |
-| `mcp-tools-list.canonical.json` | Launch 10-tool listing oracle (schemas from live capture) |
-| `mcp-tools-list.meta.json` | Endpoint, protocol version, SHA-256, exclusions |
+| `mcp-tools-list.canonical.json` | Gate-1 launch 13-tool listing oracle (schemas from live capture) |
+| `mcp-tools-list.meta.json` | Endpoint, protocol version, SHA-256, Gate-1 promotions |
 | `mcp-initialize.raw.json` / `.meta.json` | Live `initialize` (protocolVersion) |
 | `mcp-tools-call-unauth-error.raw.json` | Live unauthenticated `tools/call` error envelope |
 | `mcp-tools-call.representative.json` | Schema-derived representative success call shapes |
@@ -25,4 +25,4 @@ Immutable upstream captures for `@pome-sh/twin-gmail`. Normal tests must not req
 
 - `tools/list` and `initialize` were captured live without OAuth on 2026-07-20 against `https://gmailmcp.googleapis.com/mcp/v1`.
 - Authenticated `tools/call` success was **not** available; representative success fixtures are reconstructed from live `outputSchema` + public docs and are marked as such.
-- Live listing currently includes three tools outside the launch set (`get_message`, `apply_sensitive_thread_label`, `apply_sensitive_message_label`). The twin launch freeze is the documented 10-tool set.
+- Gate 1 expands the OSS launch set to the full live 13-tool listing, promoting `get_message`, `apply_sensitive_thread_label`, and `apply_sensitive_message_label` that Gate 0 treated as preview drift.
