+++ packages/twin-gmail/src/domain/index.ts
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+//
+// Domain barrel (F-679): re-exports only.
+export { GmailDomain } from "./gmail-domain.js";
+export type { DraftResource } from "./drafts.js";
+export type { FilterResource } from "./filters.js";
+export type { LabelResource } from "./labels.js";
